re-derivation of Schultz Feynman variational alpha=small radius
0.44 in Schultz's 1959 paper is actually 4/9=0.44444...

More importantly, the 4 pages of re-derivation + half a pencil lead led
me to realise that I was using an incorrect conversion to SI, from the
variational solutions.
You need to include a factor of the effective mass in defining 'me'.
In the case of directly using (2.4) in Schultz from the Polaron
variational parameters, this is already 'baked in' to v and w.
I regret nothing!