Basic working Hellwarth1999 with partial v,w solutions.
Need to refactor a lot of Polaronmobility monolithic function.

Perhaps if we can separate cleanly into these little partial functions,
we can reassemble the whole.
But for now, I'm not going to mess with that part of the code, just
slowly copy+paste code out as I build the new infrastructure.