Added Fourier-integral style quadrature to improve numeric stability
Still not that happy with this.
Took the form from the GSL page on this problem, after playing around
with directly using the GSL via GSL.jl. (Not much point doing this, as
even though it was super-quick, round-off errors were a bit difficult to
predict.)
I do not implement the epsilon-acceleration, just the calculation of the
bounds for the quadrature, and then hope that QuadGK.jl sorts out
everything else.

Horribly oscillatory!